+++ src/components/common/NoticeIcon.tsx
@@ -6,6 +6,7 @@ import useGetIsNotice from '@hooks/queries/useGetIsNotice';
 import { getToken, setToken } from '@utils/localStorage/token';
 import { useAppDispatch, useAppSelector } from '@features/hooks';
 import { setNotice } from '@features/noticeSlice';
+import Toast from './Toast';
 
 interface NoticeIconProps {
   isSearch?: boolean;
@@ -15,39 +16,6 @@ interface NoticeIconProps {
 const NoticeIconTag = tw.div<defaultProps>`
 flex w-[60px] justify-between 
 `;
-
-interface ToastProps {
-  setToast: (value: boolean) => void;
-  text: string;
-  [key: string]: any;
-}
-
-function Toast({ setToast, text }: ToastProps) {
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setToast(false);
-    }, 2000);
-    return () => {
-      clearTimeout(timer);
-    };
-  }, [setToast]);
-  return (
-    <div className="fixed inset-x-0 top-8 m-auto flex h-[80px] w-[300px] animate-bounce4 flex-col items-start gap-2 rounded-2xl bg-[#F5F5F5] p-3 opacity-80">
-      <div className="flex gap-2">
-        <div className="flex h-[20px] w-[20px] items-center justify-center rounded bg-[#FC6554]">
-          <Image
-            alt="logo"
-            src="/svg/icons/icon_logo_brand.svg"
-            width="15"
-            height="0"
-          />
-        </div>
-        <span>Atties</span>
-      </div>
-      <p>{text}</p>
-    </div>
-  );
-}
 
 export default React.memo(function NoticeIcon({
   isSearch = true,
@@ -77,15 +45,7 @@ export default React.memo(function NoticeIcon({
 
   return (
     <NoticeIconTag {...rest}>
-      {toast && (
-        <Toast
-          setToast={setToast}
-          text="알림이 도착하였습니다.🔔"
-          onClick={() => {
-            router.push('/notice');
-          }}
-        />
-      )}
+      {toast && <Toast setToast={setToast} text="알림이 도착하였습니다.🔔" />}
       {isSearch ? (
         <Image
           alt="search"
